Use size_t for get_data_size.\ Use uint32_t for fw_log_data timestamp. Use uint64_t for software_sensor uint64_t _unique_id. find_if() nodiscard fixed in sensor.cpp.

+++ include/librealsense2/h/rs_internal.h
@@ -394,7 +394,7 @@ const unsigned char* rs2_fw_log_message_data(rs2_firmware_log_message* msg, rs2_
 * \param[out] error     If non-null, receives any error that occurs during this call, otherwise, errors are ignored.
 * \return               size of the firmware log message data
 */
-int rs2_fw_log_message_size(rs2_firmware_log_message* msg, rs2_error** error);
+size_t rs2_fw_log_message_size(rs2_firmware_log_message* msg, rs2_error** error);
 
 
 /**
